extract several text drawing functions to graphics.ts
